fix: block component block not available yet  (#745)
During a regression test after [ctypehub/pull/743](https://github.com/galaniprojects/ctypehub/pull/743), it was found that the "Registration Block" component was showing a broken link after creating a cType from inside the app. This was because the block was still not available and the link required the block number.

Now: 

- The component will direct to the polkadot.js explorer (without querying a specific block) when block number is undefined. 
Reverted: - ~When submitting a cType creation transaction, it will be awaited until it is _InBlock_.~

## Other fix: 
Handle cTypes without a title.